[ALPS08982961] mml: fix coverity issue
[Description]
mtk-mml-test.c
19047713: avoid overflow

[Test]
Build pass. ME stress pass.

MTK-Commit-Id: 903a1acd57dd6b786d5721452d063d51df26683d

CR-Id: ALPS08982961
Feature: mml

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-test.c
@@ -583,7 +583,7 @@ static void case_general_submit_ut(struct mml_test *test,
 	ktime_get_real_ts64((struct timespec64 *)&task.end);
 	for (i = 0; i < utcfg->round; i++) {
 		timespec64_add_ns((struct timespec64 *)&task.end,
-			utcfg->interval * 1000000);
+			(u64)utcfg->interval * 1000000);
 		ret = mml_drm_submit(ctx, &task, NULL);
 		if (ret) {
 			mml_err("[test]%s submit failed: %d round: %u",
